Move couch design docs service to libs
