+++ src/commands/design/ls.ts
@@ -1,7 +1,7 @@
 import { Args, Command } from '@effect/cli';
 import { Array, Effect, Option, pipe, Record } from 'effect';
 import { initializeUrl } from '../../index';
-import { getDesignDocNames } from '../../services/couch/design-docs';
+import { getDesignDocNames } from '../../libs/couch/design-docs';
 import { getDbNames } from '../../libs/couch/dbs-info';
 
 import { logJson } from '../../libs/console';
