+++ test/services/warm-views.spec.ts
@@ -3,7 +3,7 @@ import { Effect, Layer } from 'effect';
 import { expect } from 'chai';
 import * as CouchDbsInfo from '../../src/libs/couch/dbs-info';
 import { CouchDesignInfoService } from '../../src/services/couch/design-info';
-import * as CouchDesignDocs from '../../src/services/couch/design-docs';
+import * as CouchDesignDocs from '../../src/libs/couch/design-docs';
 import { createDesignInfo } from '../utils/data-models';
 import { WarmViewsService } from '../../src/services/warm-views';
 import * as CouchDesign from '../../src/libs/couch/design';
